[CALCITE-1054] Wrong code generation for nullable Timestamp values
To fix code generation, we take a different approach to handling NULL
values. Previously when generating the Java AST we would know whether
we were intended to produce null, true or false if the value is null.

After this change, we handle null semantics by transforming the
RexNode tree, so if null values are not possible, the RexNode will
have type NOT NULL. If we have earlier checked that an expression is
not null, we can enclose it in a "CAST(... NOT NULL type)"

When unparsing, never put parentheses around function calls.

Add a variant of composeDistinctions and composeConjunctions that
obviously is not null.