[CALCITE-1384] Extension point for ALTER statements (Gabriel Reid)
Add a parser extension point for parsing of ALTER statements.

The main content of this commit is the addition of an initial
test setup to allow testing parsing extension points within
Calcite itself. This involves the addition of test-specific
parser templates and building of a test-specific parser.

Close apache/calcite#322