Start on [CALCITE-1069] In Aggregate, combine GROUPING columns into one GROUP_ID column
The plan is to deprecate the Aggregate.indicator field (strongly
encouraging people to set it to false) and to allow GROUPING and
GROUP_ID to be used as aggregate functions. They will be handled
at implement time.

With indicator = false, even queries with more than one grouping set
will just output the join keys and the aggregate functions. A join key
will be nullable if it doesn't appear in all grouping sets.

The output row type of Aggregate will be more consistent, and this will
cause fewer bugs in rules etc. in the long run.